refactor: split widget wallet integration surfaces
