+++ typescript/widgets/src/index.ts
@@ -91,100 +91,3 @@ export { useDebounce } from './utils/debounce.js';
 export { useIsSsr } from './utils/ssr.js';
 export { useInterval, useTimeout } from './utils/timeout.js';
 export { useConnectionHealthTest } from './utils/useChainConnectionTest.js';
-export {
-  AccountList,
-  AccountSummary,
-} from './walletIntegrations/AccountList.js';
-export { ConnectWalletButton } from './walletIntegrations/ConnectWalletButton.js';
-export {
-  getCosmosKitChainConfigs,
-  useCosmosAccount,
-  useCosmosActiveChain,
-  useCosmosConnectFn,
-  useCosmosDisconnectFn,
-  useCosmosTransactionFns,
-  useCosmosWalletDetails,
-  useCosmosSwitchNetwork,
-  useCosmosWatchAsset,
-} from './walletIntegrations/cosmos.js';
-export {
-  getWagmiChainConfigs,
-  useEthereumAccount,
-  useEthereumActiveChain,
-  useEthereumConnectFn,
-  useEthereumDisconnectFn,
-  useEthereumTransactionFns,
-  useEthereumWalletDetails,
-  useEthereumSwitchNetwork,
-  useEthereumWatchAsset,
-} from './walletIntegrations/ethereum.js';
-export {
-  getAccountAddressAndPubKey,
-  getAccountAddressForChain,
-  useAccountAddressForChain,
-  useAccountForChain,
-  useAccounts,
-  useActiveChains,
-  useConnectFns,
-  useDisconnectFns,
-  useTransactionFns,
-  useWalletDetails,
-  useWatchAsset,
-} from './walletIntegrations/multiProtocol.js';
-export { MultiProtocolWalletModal } from './walletIntegrations/MultiProtocolWalletModal.js';
-export {
-  useSolanaAccount,
-  useSolanaActiveChain,
-  useSolanaConnectFn,
-  useSolanaDisconnectFn,
-  useSolanaTransactionFns,
-  useSolanaWalletDetails,
-  useSolanaSwitchNetwork,
-  useSolanaWatchAsset,
-} from './walletIntegrations/solana.js';
-export {
-  getStarknetChains,
-  useStarknetAccount,
-  useStarknetActiveChain,
-  useStarknetConnectFn,
-  useStarknetDisconnectFn,
-  useStarknetTransactionFns,
-  useStarknetWalletDetails,
-  useStarknetSwitchNetwork,
-  useStarknetWatchAsset,
-} from './walletIntegrations/starknet.js';
-export {
-  useAleoAccount,
-  useAleoActiveChain,
-  useAleoConnectFn,
-  useAleoDisconnectFn,
-  useAleoTransactionFns,
-  useAleoWalletDetails,
-  useAleoSwitchNetwork,
-  useAleoWatchAsset,
-} from './walletIntegrations/aleo.js';
-export { AleoPopupProvider } from './walletIntegrations/aleo/AleoProviders.js';
-export type {
-  AccountInfo,
-  ActiveChainInfo,
-  ChainAddress,
-  ChainTransactionFns,
-  SendTransactionFn,
-  SwitchNetworkFn,
-  SwitchNetworkFns,
-  WalletDetails,
-  SendMultiTransactionFn,
-  WatchAssetFns,
-} from './walletIntegrations/types.js';
-export {
-  ethers5TxToWagmiTx,
-  findChainByRpcUrl,
-  getChainsForProtocol,
-} from './walletIntegrations/utils.js';
-export { WalletLogo } from './walletIntegrations/WalletLogo.js';
-export {
-  RdtProvider,
-  GatewayApiProvider,
-  PopupProvider,
-} from './walletIntegrations/radix/RadixProviders.js';
-export { AccountProvider } from './walletIntegrations/radix/AccountContext.js';
