+++ typescript/widgets/src/utils/useChainConnectionTest.ts
@@ -1,10 +1,8 @@
 import { useEffect, useState } from 'react';
 
-import {
-  ChainMetadata,
-  isBlockExplorerHealthy,
-  isRpcHealthy,
-} from '@hyperlane-xyz/sdk';
+import type { ChainMetadata } from '@hyperlane-xyz/sdk/metadata/chainMetadataTypes';
+import { isBlockExplorerHealthy } from '@hyperlane-xyz/sdk/providers/explorerHealthTest';
+import { isRpcHealthy } from '@hyperlane-xyz/sdk/providers/rpcHealthTest';
 import { timeout } from '@hyperlane-xyz/utils';
 
 import { ChainConnectionType } from '../chains/types.js';
