+++ typescript/widgets/src/walletIntegrations/types.ts
@@ -1,9 +1,7 @@
-import {
-  ChainName,
-  IToken,
-  TypedTransactionReceipt,
-  WarpTypedTransaction,
-} from '@hyperlane-xyz/sdk';
+import { type TypedTransactionReceipt } from '@hyperlane-xyz/sdk/providers/ProviderType';
+import type { IToken } from '@hyperlane-xyz/sdk/token/IToken';
+import type { ChainName } from '@hyperlane-xyz/sdk/types';
+import type { WarpTypedTransaction } from '@hyperlane-xyz/sdk/warp/types';
 import { HexString, ProtocolType } from '@hyperlane-xyz/utils';
 
 export interface ChainAddress {
